+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The generated wrapper struct now `#[repr(transparent)]`.
+
 ## [v0.6.1] - 2025-09-03
 
 ### Fixed